# Unified caching for `load()`, `Module()`