… crates/pcb-zen-core/src/lang/module.rs
… crates/pcb-zen/tests/snapshots/module_loading__module_nonexistent_alias.snap
… crates/pcb-zen/tests/snapshots/module_loading__module_nonexistent_file.snap
… crates/pcb/tests/snapshots/build__aggregated_warnings.snap
… crates/pcb/tests/snapshots/build__alias_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__cross_repo_unstable_warning.snap
… crates/pcb/tests/snapshots/build__default_alias_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__deny_warnings_flag.snap
… crates/pcb/tests/snapshots/build__gitlab_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__load_and_module_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__method_chaining_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__mixed_aggregated_and_unique_warnings.snap
… crates/pcb/tests/snapshots/build__mixed_stable_unstable_refs.snap
… crates/pcb/tests/snapshots/build__module_unstable_ref_warning.snap
… crates/pcb/tests/snapshots/build__transitive_unstable_dependencies.snap
… crates/pcb/tests/snapshots/build__unstable_ref_no_explicit_branch.snap
… crates/pcb/tests/snapshots/build__unstable_ref_wrong_tag.snap
… crates/pcb/tests/snapshots/build__warning_and_error_mixed.snap
… test-workspaces/with-pcb-toml/modules/LedIndicatorModule.zen