~7x speedup:

Before:
```
  Time (mean ± σ):      2.163 s ±  0.037 s    [User: 1.564 s, System: 0.585 s]
  Range (min … max):    2.110 s …  2.227 s    10 runs
```

After:
```
  Time (mean ± σ):     305.7 ms ±   2.1 ms    [User: 261.6 ms, System: 41.5 ms]
  Range (min … max):   303.2 ms … 309.3 ms    10 runs
```